Fixed some false positives with the code
Surprising absolutely nobody, it turns out that it wasn't just the test cases that were the problem.
If anyone knows what a better error than 401 is for the missing fields is, let me know. It seems trivial for now though so I'll just send it as is cuz I'm already behind schedule